+++ app/src/test/kotlin/me/xizzhu/android/joshua/progress/ReadingProgressPresenterTest.kt
@@ -16,24 +16,13 @@
 
 package me.xizzhu.android.joshua.progress
 
-import android.view.View
-import android.widget.ProgressBar
-import kotlinx.coroutines.flow.emptyFlow
-import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.test.runBlockingTest
+import androidx.lifecycle.LifecycleCoroutineScope
 import me.xizzhu.android.joshua.Navigator
 import me.xizzhu.android.joshua.core.Bible
 import me.xizzhu.android.joshua.core.ReadingProgress
-import me.xizzhu.android.joshua.core.Settings
-import me.xizzhu.android.joshua.core.VerseIndex
-import me.xizzhu.android.joshua.infra.arch.ViewData
 import me.xizzhu.android.joshua.tests.BaseUnitTest
-import me.xizzhu.android.joshua.ui.fadeIn
 import me.xizzhu.android.joshua.ui.recyclerview.BaseItem
-import me.xizzhu.android.joshua.ui.recyclerview.CommonRecyclerView
 import org.mockito.Mock
-import org.mockito.Mockito.*
 import kotlin.test.*
 
 class ReadingProgressPresenterTest : BaseUnitTest() {
@@ -42,91 +31,17 @@ class ReadingProgressPresenterTest : BaseUnitTest() {
     @Mock
     private lateinit var navigator: Navigator
     @Mock
-    private lateinit var readingProgressInteractor: ReadingProgressInteractor
+    private lateinit var readingProgressViewModel: ReadingProgressViewModel
     @Mock
-    private lateinit var loadingSpinner: ProgressBar
-    @Mock
-    private lateinit var readingProgressListView: CommonRecyclerView
+    private lateinit var lifecycleCoroutineScope: LifecycleCoroutineScope
 
-    private lateinit var readingProgressViewHolder: ReadingProgressViewHolder
     private lateinit var readingProgressPresenter: ReadingProgressPresenter
 
     @BeforeTest
     override fun setup() {
         super.setup()
 
-        runBlocking {
-            `when`(readingProgressInteractor.settings()).thenReturn(emptyFlow())
-            `when`(readingProgressInteractor.bookNames()).thenReturn(List(Bible.BOOK_COUNT) { i -> i.toString() })
-            `when`(readingProgressInteractor.readingProgress()).thenReturn(ReadingProgress(0, 0L, emptyList()))
-
-            readingProgressViewHolder = ReadingProgressViewHolder(loadingSpinner, readingProgressListView)
-            readingProgressPresenter = ReadingProgressPresenter(readingProgressActivity, navigator, readingProgressInteractor, testDispatcher)
-        }
-    }
-
-    @Test
-    fun testObserveSettings() = testDispatcher.runBlockingTest {
-        val settings = Settings(false, true, 1, true, true)
-        `when`(readingProgressInteractor.settings()).thenReturn(flowOf(ViewData.loading(), ViewData.success(settings), ViewData.error()))
-
-        readingProgressPresenter.create(readingProgressViewHolder)
-        verify(readingProgressListView, times(1)).setSettings(settings)
-        verify(readingProgressListView, never()).setSettings(Settings.DEFAULT)
-
-        readingProgressPresenter.destroy()
-    }
-
-    @Test
-    fun testLoadReadingProgress() = testDispatcher.runBlockingTest {
-        `when`(readingProgressInteractor.readingProgress()).thenReturn(ReadingProgress(5, 4321L, emptyList()))
-
-        readingProgressPresenter.create(readingProgressViewHolder)
-        readingProgressPresenter.start()
-
-        with(inOrder(loadingSpinner, readingProgressListView)) {
-            verify(loadingSpinner, times(1)).fadeIn()
-            verify(readingProgressListView, times(1)).visibility = View.GONE
-            verify(readingProgressListView, times(1)).setItems(any())
-            verify(readingProgressListView, times(1)).fadeIn()
-            verify(loadingSpinner, times(1)).visibility = View.GONE
-        }
-
-        readingProgressPresenter.stop()
-        readingProgressPresenter.destroy()
-    }
-
-    @Test
-    fun testLoadReadingProgressWithBookNamesException() = testDispatcher.runBlockingTest {
-        val exception = RuntimeException("Random exception")
-        `when`(readingProgressInteractor.bookNames()).thenThrow(exception)
-
-        readingProgressPresenter.create(readingProgressViewHolder)
-        readingProgressPresenter.start()
-
-        with(inOrder(loadingSpinner, readingProgressListView)) {
-            verify(loadingSpinner, times(1)).fadeIn()
-            verify(readingProgressListView, times(1)).visibility = View.GONE
-            verify(loadingSpinner, times(1)).visibility = View.GONE
-        }
-        verify(readingProgressListView, never()).setItems(any())
-        verify(readingProgressListView, never()).fadeIn()
-
-        readingProgressPresenter.stop()
-        readingProgressPresenter.destroy()
-    }
-
-    @Test
-    fun testOpenChapter() = testDispatcher.runBlockingTest {
-        readingProgressPresenter.create(readingProgressViewHolder)
-
-        val bookIndex = 1
-        val chapterIndex = 2
-        readingProgressPresenter.openChapter(bookIndex, chapterIndex)
-        verify(readingProgressInteractor, times(1)).saveCurrentVerseIndex(VerseIndex(bookIndex, chapterIndex, 0))
-        verify(navigator, times(1)).navigate(readingProgressActivity, Navigator.SCREEN_READING)
-
-        readingProgressPresenter.destroy()
+        readingProgressPresenter = ReadingProgressPresenter(readingProgressActivity, navigator, readingProgressViewModel, lifecycleCoroutineScope)
     }
 
     @Test
@@ -170,14 +85,5 @@ class ReadingProgressPresenterTest : BaseUnitTest() {
                 }
         val actual = with(readingProgressPresenter) { readingProgress.toItems(Array(Bible.BOOK_COUNT) { "bookName$it" }.toList()) }
         assertEquals(expected, actual)
-    }
-
-    @Test
-    fun testOpenChapterWithException() = testDispatcher.runBlockingTest {
-        `when`(readingProgressInteractor.saveCurrentVerseIndex(any())).thenThrow(RuntimeException("Random exception"))
-
-        readingProgressPresenter.create(readingProgressViewHolder)
-        readingProgressPresenter.openChapter(0, 0)
-        readingProgressPresenter.destroy()
     }
 }