+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
 - Refactoring
   - Used ViewPager2 instead of ViewPager
 - Updated dependencies:
+  - Kotlin JVM target to 1.8
   - Gradle to 6.1.1
   - Android build tools to 29.0.3
   - AndroidX Core to 1.2.0, Material to 1.1.0, ViewPager2 to 1.0.0